Arrange states in a semantic way
There are workflow categories and there are utility categories

In progress could mean submitted but not offered. It's a categorisation
of states WRT the progression of the application.

Reapply states are contextual and serve a purpose in a specific context.
A utility category could contain states across the workflow categories.